Review notifications snooze handling
Cannot rely on setTimeout, especially in serviceWorker, as the service
worker can suspend any time.
Not sure about having a timeout in the app either...
Instead, store in the Redux state the time when the snooze button is
pressed and check if the snooze timeout is expired each time data comes and the
notifications should trigger.
This means the paused state in the header icon is not changing after the
specified timeout, but will change the next time data comes (which in
the worst case is 1 minute after).